Added fog
Removed that cube and made a collision for that wall that was missing
one.
Added emissive to the lights, that are still using a standard material.
They turn super duper bright with toonlit and lighting will need to
change for it to look normal. I think we should all be there if we
decide to change that. Also the wall light prefab has 3 lights in it.