New sun beam particle
Messed around with making new sun shafts. Tried to see how a panning
texture would look, I think particle may be the way to go. Still have
the script in there.
New sun shaft particle is a prefab in the playground and temple, the old
asset creator one was set to inactive.
The transparency of the particle messes with the transparency of the
moss behind it a little bit.