Nyeh ha ha
Imported click clock wood from Banjo Kazooie, also in a self-contained
folder so we can get rid of it whenver.
There was an incident where jumping on one of the higher walkways ported
Pyerus to the ground and I couldn't replicate it.
Also that's how many materials are there holy cow.